+++ docs/conf.py
@@ -29,7 +29,7 @@ copyright = "2021, TwitchIO"
 author = "TwitchIO"
 
 # The full version, including alpha/beta/rc tags
-release = "2.1.4"
+release = "2.1.5"
 
 
 # -- General configuration ---------------------------------------------------