+++ twitchio/ext/pubsub/models.py
@@ -280,11 +280,9 @@ class PubSubModerationActionBanRequest(PubSubMessage):
         The user that created the action.
     target: :class:`twitchio.PartialUser`
         The target of this action.
-    from_automod: :class:`bool`
-        Whether this action was done automatically or not.
     """
 
-    __slots__ = "action", "args", "created_by", "message_id", "target", "from_automod"
+    __slots__ = "action", "args", "created_by", "message_id", "target"
 
     def __init__(self, client: Client, topic: str, data: dict):
         super().__init__(client, topic, data)
@@ -320,7 +318,7 @@ class PubSubModerationActionChannelTerms(PubSubMessage):
         The requester of this Term.
     """
 
-    __slots__ = "type", "channel_id", "id", "text", "requester", "expires_at", "updated_at", "from_automod"
+    __slots__ = "type", "channel_id", "id", "text", "requester", "expires_at", "updated_at"
 
     def __init__(self, client: Client, topic: str, data: dict):
         super().__init__(client, topic, data)
@@ -332,12 +330,12 @@ class PubSubModerationActionChannelTerms(PubSubMessage):
             client._http, data["message"]["data"]["requester_id"], data["message"]["data"]["requester_login"]
         )
 
-        self.expires_at = self.updated_at = None
-        if data["message"]["data"]["expires_at"]:
-            self.expires_at = parse_timestamp(data["message"]["data"]["expires_at"])
-
-        if data["message"]["data"]["updated_at"]:
-            self.updated_at = parse_timestamp(data["message"]["data"]["expires_at"])
+        self.expires_at = (
+            parse_timestamp(data["message"]["data"]["expires_at"]) if data["message"]["data"]["expires_at"] else None
+        )
+        self.updated_at = (
+            parse_timestamp(data["message"]["data"]["updated_at"]) if data["message"]["data"]["updated_at"] else None
+        )
 
 
 class PubSubModerationActionModeratorAdd(PubSubMessage):